Castle trap doors fix
     Recently discussed in the newsgroup, and it sounded familiar so may
have been reported directly earlier:

  You fly through the trap door.  You fly down along the stairs.

when polymorphed into a flying creature or riding a flying steed.  It
only happens at the castle, and only happens because the go-down code is
explicitly setting the arrival point to be the Valley stairs.  That's how
the castle trap doors used to work as traps too, but they were changed to
dump you randomly near the stairs quite some time ago.  Fix it by making
intentional triggering work the same now; this also prevents the false
feedback (even if you happen to arrive on the stairs by coincidence).