temple sounds (trunk only)
     The user (<email deleted>) who recently suggested a
dump command for containers also wanted atmospheric sounds on levels which
have altars.  Right now we'd have to find unattended altars the hard way
(by scanning the entire level) but we could add a counter (or set of
counters, one per alignment) like for fountains and sinks if we really
wanted to do that.  [Now that I think about it, the #overview patch may
have already done something of the sort.]  But what noises would an altar
be expected to produce?  This only adds sounds for temples, where the
attending priest can be the source of the noise.

     I'm not real thrilled with the initial set of sounds, particularly
the hallucinating one, but the implementation works.  The "carcass" one is
a little clumsy; it's intended to add a hint for new players who haven't
figured out what the #offer command does.